task done

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,8 +9,8 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int temp = n ^ m;
-	int i = 0;
+	unsigned long int temp = n ^ m;
+	unsigned int i = 0;
 
 	while (temp)
 	{
